[ft-#96-middleware] fix responses in tests; add members.testFn for renaming tests
