+++ tests/unit/http/response_test.ts
@@ -14,8 +14,18 @@ let responseFormatted;
 
 members.test(function Response_generateResponse_json_default() {
   members.assert.equal(
-    response.generateResponse(),
-    `{"status_code":200,"status_message":"OK","request":{"url":"/","method":"GET"},"body":"This is my body"}`
+    JSON.parse(response.generateResponse()),
+    {
+      status_code: 200,
+      status_message: "OK",
+      body: "This is my body",
+      request: {
+        method: "GET",
+        uri: "/",
+        url_query_params: {},
+        url: "/"
+      }
+    }
   );
 });
 
@@ -34,6 +44,12 @@ members.test(function Response_generateResponse_xml() {
   <statuscode>200</statuscode>
   <statusmessage>OK</statusmessage>
   <body>This is my body</body>
+  <request>
+    <method>GET</method>
+    <uri>/</uri>
+    <url_query_params>{}</url_query_params>
+    <url>/</url>
+  </request>
 </response>`
   );
   response.headers.set("Content-Type", "application/xml");
@@ -44,6 +60,12 @@ members.test(function Response_generateResponse_xml() {
   <statuscode>200</statuscode>
   <statusmessage>OK</statusmessage>
   <body>This is my body</body>
+  <request>
+    <method>GET</method>
+    <uri>/</uri>
+    <url_query_params>{}</url_query_params>
+    <url>/</url>
+  </request>
 </response>`
   );
 });
